Added range functionality to process multiple radsecproxy log files
Previous script could only process a single file each time it executed.

Added an argument parser which allows users to select a period of logs
to analyse by providing a start and end date. If no dates are provided,
it will default to the previous day's log.

Added valid_date function to provide checking of inputted start and end
dates.